+++ adapters/Unity/UnityAdapter/Categories/NSError+Unity.h
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2021 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,13 +15,12 @@
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface GADMAdapterVungleNativeAd : NSObject
+NS_ASSUME_NONNULL_BEGIN
 
-/// Asks the receiver to render the ad configuration.
-- (nonnull instancetype)initNativeAdForAdConfiguration:(nonnull GADMediationNativeAdConfiguration *)adConfiguration
-                                     completionHandler:(nonnull GADMediationNativeLoadCompletionHandler)completionHandler;
-- (nonnull instancetype)init NS_UNAVAILABLE;
-- (void)requestAd;
-
+@interface NSError (Unity)
++ (NSError *) noValidGameId;
++ (NSError *) unsupportedBannerGADAdSize:(GADAdSize)adSize;
++ (NSError *) adNotAvailablePerPlacement:(NSString*)placementId;
 @end
 
+NS_ASSUME_NONNULL_END